Add support to declare multiple css/js per module with loading condition:
A css/js can be declared to only load if the module in question is the
page module being rendered, or to always be loaded even if the module
is just a dependency. The second case is useful for plugin type
modules. See one of the config.json files updated in this commit on
how to specify css/js to load. true = always load, false = only if page
module.