index.php: Hacky way of telling the php error handler to skip stack frames
This can be used to show more meaningful error messages in debug mode at
the bottom of the page. Prefix the error message with [skip:N] where
N is the number of stack frames to skip when printing the error message.
The php error log will still receive the complete stack trace.