[sysconfig] Stuff ssh config into "empty" global config
This is an ugly hack. The "no sysconfig" runmode option
creates an obvious problem if an ssh config exists: It
won't be provisioned to the client, so it's not reachable
via ssh. We need a proper mechanism for managing config,
that flags modules by whether they should be ignored for
"no sysconfig" runmode or not, and while we're at it,
make it possible to assign additional modules to rooms.