[locationinfo] Don't refresh calendars indefinitely
We try to refresh calendars that have recently been used
to coalesce requests, which helps especially with HisInOne.
However, this would in turn make the calendar recently
refreshed so it would keep going on and on, as long as there
is at least one active info panel making actual requests.